@@ -1,6 +1,6 @@
 {
   "name": "@oruga-ui/docs-next",
-  "version": "0.9.1",
+  "version": "0.9.2",
   "homepage": "https://oruga-ui.com",
   "description": "UI components for Vue.js and CSS framework agnostic",
   "author": "Walter Tommasi <tommsi20@gmail.com>",
@@ -9,15 +9,17 @@
   "private": true,
   "scripts": {
     "dev": "vitepress dev",
-    "docs:dev": "concurrently \"vue-docgen --watch\" \"vitepress dev\"",
-    "docs:gen": "vue-docgen",
-    "docs:build": "vitepress build",
+    "build": "vitepress build",
+    "docs:dev": "concurrently \"npm run docs:gen:watch\" \"npm run dev\"",
+    "docs:build": "npm run build",
+    "docs:gen": "tsx src/docgen.ts",
+    "docs:gen:watch": "tsx src/docgen.ts --watch",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix",
     "update": "ncu -u"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.6.0",
-    "@fortawesome/free-solid-svg-icons": "^6.6.0",
+    "@fortawesome/fontawesome-svg-core": "^6.7.0",
+    "@fortawesome/free-solid-svg-icons": "^6.7.0",
     "@fortawesome/vue-fontawesome": "^3.0.8",
     "@highlightjs/vue-plugin": "2.1.2",
     "@oruga-ui/theme-bootstrap": "0.7.1",
@@ -28,19 +30,20 @@
     "markdown-it-highlightjs": "^4.2.0"
   },
   "devDependencies": {
-    "@types/node": "^20.14.10",
+    "@types/node": "^22.9.1",
     "@vue/eslint-config-prettier": "^10.1.0",
     "@vue/eslint-config-typescript": "^14.1.3",
-    "@vue/tsconfig": "^0.5.1",
-    "concurrently": "9.0.1",
-    "eslint": "^9.14.0",
-    "eslint-plugin-vue": "^9.30.0",
-    "npm-check-updates": "^17.1.10",
+    "@vue/tsconfig": "^0.6.0",
+    "concurrently": "9.1.0",
+    "eslint": "^9.15.0",
+    "eslint-plugin-vue": "^9.31.0",
+    "npm-check-updates": "^17.1.11",
     "prettier": "^3.3.3",
-    "sass": "^1.80.6",
+    "sass": "^1.81.0",
+    "tsx": "^4.19.2",
     "typescript": "5.6.3",
     "vitepress": "^1.5.0",
-    "vue": "^3.5.12",
+    "vue": "^3.5.13",
     "vue-component-meta": "^2.1.10",
     "vue-docgen-cli": "4.79.0",
     "vue-tsc": "^2.1.10"
