@@ -1,6 +1,6 @@
 {
   "name": "@oruga-ui/oruga-next",
-  "version": "0.9.1",
+  "version": "0.9.2",
   "homepage": "https://oruga-ui.com",
   "description": "UI components for Vue.js and CSS framework agnostic",
   "author": "Walter Tommasi <tommsi20@gmail.com>",
@@ -75,32 +75,32 @@
     "@babel/core": "7.26.0",
     "@babel/preset-env": "7.26.0",
     "@rollup/plugin-babel": "6.0.4",
-    "@tsconfig/node20": "^20.1.4",
+    "@tsconfig/node22": "^22.0.0",
     "@types/jest-axe": "^3.5.9",
     "@types/jsdom": "^21.1.7",
-    "@types/node": "^20.14.10",
-    "@vitejs/plugin-vue": "^5.1.4",
-    "@vitest/coverage-istanbul": "^2.1.4",
+    "@types/node": "^22.9.0",
+    "@vitejs/plugin-vue": "^5.2.0",
+    "@vitest/coverage-istanbul": "^2.1.5",
     "@vue/eslint-config-prettier": "^10.1.0",
     "@vue/eslint-config-typescript": "^14.1.3",
     "@vue/test-utils": "^2.4.6",
-    "@vue/tsconfig": "^0.5.1",
+    "@vue/tsconfig": "^0.6.0",
     "core-js": "3.39.0",
-    "eslint": "^9.14.0",
+    "eslint": "^9.15.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-vue": "^9.30.0",
+    "eslint-plugin-vue": "^9.31.0",
     "eslint-plugin-vuejs-accessibility": "^2.4.1",
     "jest-axe": "^9.0.0",
     "jsdom": "^25.0.1",
-    "npm-check-updates": "^17.1.9",
+    "npm-check-updates": "^17.1.11",
     "prettier": "^3.3.3",
     "rimraf": "^6.0.1",
     "typescript": "^5.6.3",
-    "vite": "^5.4.10",
+    "vite": "^5.4.11",
     "vite-plugin-dts": "^4.3.0",
-    "vite-tsconfig-paths": "^5.0.1",
-    "vitest": "^2.1.4",
-    "vue": "^3.5.12",
+    "vite-tsconfig-paths": "^5.1.2",
+    "vitest": "^2.1.5",
+    "vue": "^3.5.13",
     "vue-component-meta": "^2.1.10",
     "vue-tsc": "^2.1.10"
   }
