@@ -1,4 +1,4 @@
-const replace = require('replace-in-file')
+import replace from 'replace-in-file';
 
 const jsonPaths = [
   'package-lock.json',
@@ -10,7 +10,7 @@ const version = require('../package.json').version
 
 const configJsonOptions = [{
   files: jsonPaths,
-  from: /"name": "@oruga-ui\/(.*)",([^"]*)"version": .+/ig, 
+  from: /"name": "@oruga-ui\/(.*)",([^"]*)"version": .+/ig,
   to: (match) => match.replace(/"version": .+/i, `"version": "${version}",`),
 },
 {
@@ -19,13 +19,13 @@ const configJsonOptions = [{
   to: (match) => match.replace(/oruga-next": "(.*)",/i, `oruga-next": "${version}",`),
 }]
 
-const replaceInFile = (config) => 
-   replace.sync(config).map((el) => el.file);
+const replaceInFile = (config) =>
+  replace.sync(config).map((el) => el.file);
 
 try {
   configJsonOptions.forEach(options => replaceInFile(options));
   console.info('Modified files:', jsonPaths.join(', '));
 } catch (error) {
-  console.error(e);
+  console.error(error);
   process.exit(1);
 }
