@@ -77,7 +77,7 @@ export default defineConfig({
             { text: "Expo", link: "/expo/" },
             {
                 text: "Support",
-                items: [{ text: "Ko-fi", link: "https://ko-fi.com/orugaui" }],
+                items: [{ text: "Ko-fi", link: "https://ko-fi.com/mlmoravek" }],
             },
             {
                 text: `v${version}`,
@@ -138,6 +138,8 @@ export default defineConfig({
                 "@": fileURLToPath(
                     new URL("./../../oruga/src", import.meta.url),
                 ),
+                // add '@docs' alias to docs src folder
+                "@docs": fileURLToPath(new URL("./../src", import.meta.url)),
             },
         },
     },
