@@ -6,8 +6,8 @@ const isFetching = ref(false);
 const page = ref(1);
 const totalPages = ref(1);
 
-const options = ref<OptionsPropItem[]>([]);
-const selected = ref<object>();
+const options = ref<OptionsPropItem<any>[]>([]);
+const selected = ref<any>();
 const value = ref("");
 
 async function getAsyncData(_value): Promise<void> {
