@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import { computed, type PropType } from "vue";
+import { computed } from "vue";
 
 import { getDefault } from "@/utils/config";
 import { defineClasses, useProviderChild } from "@/composables";
 
 import type { CarouselComponent } from "./types";
-import type { ComponentClass } from "@/types";
+import type { CarouselItemProps } from "./props";
 
 /**
  * A Slideshow item used by the carousel
@@ -17,35 +17,13 @@ defineOptions({
     configField: "carousel",
 });
 
-const props = defineProps({
-    /** Override existing theme classes completely */
-    override: { type: Boolean, default: undefined },
-    /** Make item clickable */
-    clickable: { type: Boolean, default: false },
-    /** Role attribute to be passed to the div wrapper for better accessibility */
-    ariaRole: {
-        type: String,
-        default: () => getDefault("carousel.ariaRole", "option"),
-    },
-    // class props (will not be displayed in the docs)
-    /** Class of carousel item */
-    itemClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of carousel item when is active */
-    itemActiveClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of carousel item when is clickable */
-    itemClickableClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
+const props = withDefaults(defineProps<CarouselItemProps>(), {
+    override: undefined,
+    clickable: false,
+    ariaRole: () => getDefault("carousel.ariaRole", "option"),
 });
 
-// Inject functionalities and data from the parent carousel component
+/** inject functionalities and data from the parent component */
 const { parent, item } = useProviderChild<CarouselComponent>();
 
 const isActive = computed(() => parent.value.activeIndex === item.value.index);
@@ -76,8 +54,8 @@ const itemClasses = defineClasses(
         v-if="parent"
         :class="itemClasses"
         :style="itemStyle"
-        :data-id="`carousel-${item.identifier}`"
         data-oruga="carousel-item"
+        :data-id="`carousel-${item.identifier}`"
         :role="ariaRole"
         aria-roledescription="item"
         :aria-selected="isActive"
