@@ -1,5 +1,5 @@
 <script setup lang="ts" generic="T">
-import { computed, useAttrs, useId, useTemplateRef } from "vue";
+import { computed, useAttrs, useId, useSlots, useTemplateRef } from "vue";
 
 import { getDefault } from "@/utils/config";
 import { defineClasses, useInputHandler } from "@/composables";
@@ -23,18 +23,18 @@ defineOptions({
 const props = withDefaults(defineProps<CheckboxProps<T>>(), {
     override: undefined,
     modelValue: undefined,
+    id: () => useId(),
     variant: () => getDefault("checkbox.variant"),
     size: () => getDefault("checkbox.size"),
     label: undefined,
     indeterminate: false,
-    nativeValue: undefined,
-    disabled: false,
     required: false,
+    disabled: false,
     name: undefined,
+    nativeValue: undefined,
     trueValue: undefined,
     falseValue: undefined,
     autocomplete: () => getDefault("checkbox.autocomplete", "off"),
-    id: () => useId(),
     useHtml5Validation: () => getDefault("useHtml5Validation", true),
     customValidity: "",
 });
@@ -51,11 +51,6 @@ const emits = defineEmits<{
      * @param event {Event} native event
      */
     input: [value: T | T[], event: Event];
-    /**
-     * indeterminate prop two-way binding
-     * @param value {boolean} updated indeterminate prop
-     */
-    "update:indeterminate": [value: boolean];
     /**
      * on input focus event
      * @param event {Event} native event
@@ -85,23 +80,20 @@ const { onBlur, onFocus, onInvalid, setFocus } = useInputHandler(
 // inject parent field component if used inside one
 const { parentField } = injectField();
 
+// set field labelId or create a unique label id if a label is given
+const labelId =
+    !!parentField.value || !!props.label || !!useSlots().label
+        ? parentField.value?.labelId || useId()
+        : undefined;
+
+// if no `label` is given and `id` is given set as `for` property on o-field wrapper
+if (!props.label && props.id) parentField.value?.setInputId(props.id);
+
 const vmodel = defineModel<T | T[]>({ default: undefined });
-
-// if not `label` is given and `id` is given set as `for` property on o-field wrapper
-if (!props.label && props.id) parentField?.value?.setInputId(props.id);
-
-const isIndeterminate = defineModel<boolean>("indeterminate", {
-    default: false,
-});
-
-const _trueValue =
-    typeof props.trueValue === "undefined" ? true : props.trueValue;
-const _falseValue =
-    typeof props.falseValue === "undefined" ? false : props.falseValue;
 
 const isChecked = computed(
     () =>
-        vmodel.value === _trueValue ||
+        vmodel.value === (props.trueValue ?? true) ||
         (Array.isArray(vmodel.value) &&
             vmodel.value.includes(props.nativeValue as T)),
 );
@@ -115,7 +107,7 @@ function onInput(event: Event): void {
 const attrs = useAttrs();
 
 const inputBind = computed(() => ({
-    ...parentField?.value?.inputAttrs,
+    ...parentField.value?.inputAttrs,
     ...attrs,
 }));
 
@@ -144,7 +136,7 @@ const inputClasses = defineClasses(
         "indeterminateClass",
         "o-chk__input--indeterminate",
         null,
-        isIndeterminate,
+        computed(() => props.indeterminate),
     ],
 );
 
@@ -157,14 +149,7 @@ defineExpose({ focus: setFocus, value: vmodel });
 </script>
 
 <template>
-    <label
-        ref="label"
-        :class="rootClasses"
-        data-oruga="checkbox"
-        role="checkbox"
-        :aria-checked="isChecked"
-        @click.stop="setFocus"
-        @keydown.prevent.enter="setFocus">
+    <div :class="rootClasses" data-oruga="checkbox">
         <input
             v-bind="inputBind"
             :id="id"
@@ -173,25 +158,30 @@ defineExpose({ focus: setFocus, value: vmodel });
             type="checkbox"
             data-oruga-input="checkbox"
             :class="inputClasses"
-            :disabled="disabled"
-            :required="required"
             :name="name"
-            :autocomplete="autocomplete"
             :value="nativeValue"
-            :indeterminate.prop="indeterminate"
-            :true-value="_trueValue"
-            :false-value="_falseValue"
-            @click.stop
+            :true-value="trueValue ?? true"
+            :false-value="falseValue ?? false"
+            :required="required"
+            :indeterminate="indeterminate"
+            :disabled="disabled"
+            :autocomplete="autocomplete"
+            :aria-checked="indeterminate ? 'mixed' : isChecked"
+            :aria-labelledby="labelId"
             @blur="onBlur"
             @focus="onFocus"
             @invalid="onInvalid"
-            @input="onInput" />
+            @change="onInput" />
 
-        <span v-if="label || $slots.default" :class="labelClasses">
+        <label
+            v-if="label || $slots.default"
+            :id="labelId"
+            :for="id"
+            :class="labelClasses">
             <!--
                 @slot Content slot, default is label prop
             -->
             <slot>{{ label }}</slot>
-        </span>
-    </label>
+        </label>
+    </div>
 </template>
