@@ -1,10 +1,10 @@
 <script setup lang="ts">
-import { useId, type PropType } from "vue";
+import { useId } from "vue";
 
 import { getDefault } from "@/utils/config";
 import { defineClasses } from "@/composables";
 
-import type { ComponentClass } from "@/types";
+import type { CollapseProps } from "./props";
 
 /**
  * An easy way to toggle what you want
@@ -17,43 +17,12 @@ defineOptions({
     configField: "collapse",
 });
 
-defineProps({
-    /** Override existing theme classes completely */
-    override: { type: Boolean, default: undefined },
-    /** Whether collapse is open or not, use v-model:open to make it two-way binding */
-    open: { type: Boolean, default: true },
-    /** Custom animation (transition name) */
-    animation: {
-        type: String,
-        default: () => getDefault("collapse.animation", "fade"),
-    },
-    /** Id property of the content container - default is an uuid */
-    contentId: { type: String, default: () => useId() },
-    /**
-     * Trigger position
-     * @values top, bottom
-     */
-    position: {
-        type: String as PropType<"top" | "bottom">,
-        default: () => getDefault("collapse.position", "top"),
-        validator: (value: string) => ["top", "bottom"].includes(value),
-    },
-    // class props (will not be displayed in the docs)
-    /** Class of the root element */
-    rootClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the trigger element */
-    triggerClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the content */
-    contentClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
+withDefaults(defineProps<CollapseProps>(), {
+    override: undefined,
+    open: true,
+    animation: () => getDefault("collapse.animation", "fade"),
+    contentId: () => useId(),
+    position: () => getDefault("collapse.position", "top"),
 });
 
 const emits = defineEmits<{
