@@ -6,7 +6,6 @@ import {
     watch,
     useId,
     useTemplateRef,
-    type PropType,
     type VNodeArrayChildren,
 } from "vue";
 
@@ -16,7 +15,7 @@ import { defineClasses, useMatchMedia } from "@/composables";
 
 import { injectField, provideField } from "./fieldInjection";
 
-import type { ComponentClass, DynamicComponent } from "@/types";
+import type { FieldProps } from "./props";
 
 /**
  * Fields are used to add functionality to controls and to attach/group components and elements together
@@ -29,180 +28,70 @@ defineOptions({
     configField: "field",
 });
 
-const props = defineProps({
-    /** Override existing theme classes completely */
-    override: { type: Boolean, default: undefined },
-    /**
-     * Color of the field and help message, also adds a matching icon.
-     * Used by Input, Select and Autocomplete.
-     * @values primary, info, success, warning, danger, and any other custom color
-     */
-    variant: { type: String, default: undefined },
-    /** Field label */
-    label: { type: String, default: undefined },
-    /**
-     * Vertical size of input
-     * @values small, medium, large
-     */
-    labelSize: {
-        type: String,
-        default: () => getDefault("field.labelsize"),
-    },
-    /** Same as native `for` set on the label */
-    labelFor: { type: String, default: undefined },
-    /** Help message text */
-    message: { type: String, default: undefined },
-    messageTag: {
-        type: [String, Object, Function] as PropType<DynamicComponent>,
-        default: () => getDefault("field.messageTag", "p"),
-    },
-    /**
-     * Direct child components/elements of Field will be grouped horizontally
-     * (see which ones at the top of the page).
-     */
-    grouped: { type: Boolean, default: false },
-    /** Allow controls to fill up multiple lines, making it responsive */
-    groupMultiline: { type: Boolean, default: false },
-    /** Group label and control on the same line for horizontal forms */
-    horizontal: { type: Boolean, default: false },
-    /** Field automatically attach controls together */
-    addons: { type: Boolean, default: false },
-    /** Mobile breakpoint as `max-width` value */
-    mobileBreakpoint: {
-        type: String,
-        default: () => getDefault("field.mobileBreakpoint"),
-    },
-    // class props (will not be displayed in the docs)
-    /** Class of the root element */
-    rootClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class to align label and control in horizontal forms */
-    horizontalClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class for field label when horizontal */
-    horizontalLabelClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** "Class for field body when horizontal */
-    horizontalBodyClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class when fields are grouped together */
-    groupedClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class when fields fill up multiple lines */
-    groupMultilineClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class for field label */
-    labelClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class for field label size */
-    labelSizeClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the label field variant */
-    labelVariantClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class for field body */
-    bodyClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class for components automatically attached together when inside a field */
-    addonsClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class for the field message */
-    messageClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the message field variant */
-    messageVariantClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of file component when on mobile */
-    mobileClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class for the focused field */
-    focusedClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class for the filled field */
-    filledClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
+const props = withDefaults(defineProps<FieldProps>(), {
+    override: undefined,
+    variant: undefined,
+    label: undefined,
+    labelSize: () => getDefault("field.labelsize"),
+    labelFor: undefined,
+    message: undefined,
+    messageTag: () => getDefault("field.messageTag", "p"),
+    grouped: false,
+    groupMultiline: false,
+    horizontal: false,
+    addons: false,
+    mobileBreakpoint: () => getDefault("field.mobileBreakpoint"),
 });
 
 const { isMobile } = useMatchMedia(props.mobileBreakpoint);
 
+/** a unique id for the field message to associate an input with */
+const messageId = useId();
+
+/** a unique id for the field label to associate an input with */
+const labelId = useId();
+
+/** the unique id for the input to associate the label with */
 const inputId = ref(props.labelFor);
 watch(
     () => props.labelFor,
     (v) => (inputId.value = v),
 );
 
-/** Set internal variant when prop change. */
+/** set internal variant when prop change */
 const fieldVariant = ref(props.variant);
 watch(
     () => props.variant,
     (v) => (fieldVariant.value = v),
 );
 
-/** Set internal message when prop change. */
+/** set internal message when prop change */
 const fieldMessage = ref(props.message);
 watch(
     () => props.message,
     (v) => (fieldMessage.value = v),
 );
 
-/** Set parent message if we use Field in Field. */
-watch(
-    () => fieldMessage.value,
-    (value) => {
-        if (parentField?.value?.hasInnerField) {
-            if (!parentField.value.variant)
-                parentField.value.setVariant(fieldVariant.value);
-            if (!parentField.value.message) parentField.value.setMessage(value);
-        }
-    },
-);
+/** set parent message if we use Field in Field */
+watch(fieldMessage, (value) => {
+    if (parentField.value && parentField.value.hasInnerField) {
+        if (!parentField.value.variant)
+            parentField.value.setVariant(fieldVariant.value);
+        if (!parentField.value.message) parentField.value.setMessage(value);
+    }
+});
 
-/** a uniqe id for the message slot to associate an input to the field message */
-const messageId = useId();
-
-/** this can be set from outside to update the focus state */
+/** this can be set from inputs to update the focus state */
 const isFocused = ref(false);
-/** this can be set from outside to update the filled state */
+/** this can be set from inputs to update the filled state */
 const isFilled = ref(false);
 /** this can be set from sub fields to update the has inner field state */
-const hasInnerField = ref<boolean>(false);
+const hasInnerField = ref(false);
 
 // inject parent field component if used inside one
 const { parentField } = injectField();
 // tell parent field it has an inner field
-if (parentField?.value) parentField.value.addInnerField();
+if (parentField.value) parentField.value.addInnerField();
 
 const slots = useSlots();
 
@@ -252,11 +141,12 @@ function setInputId(value: string): void {
     inputId.value = value;
 }
 
-const inputAttrs = computed(() =>
-    fieldVariant.value === "error"
+const inputAttrs = computed(() => ({
+    "aria-labelledby": labelId,
+    ...(fieldVariant.value === "error"
         ? { "aria-errormessage": messageId }
-        : { "aria-describedby": messageId },
-);
+        : { "aria-describedby": messageId }),
+}));
 
 // Provided data is a computed ref to enjure reactivity.
 const provideData = computed(() => ({
@@ -265,6 +155,7 @@ const provideData = computed(() => ({
     hasInnerField: hasInnerField.value,
     variant: fieldVariant.value,
     message: fieldMessage.value,
+    labelId,
     inputAttrs: inputAttrs.value,
     addInnerField,
     setInputId,
@@ -351,7 +242,11 @@ const innerFieldClasses = defineClasses(
 <template>
     <div ref="rootElement" data-oruga="field" :class="rootClasses">
         <div v-if="horizontal" :class="horizontalLabelClasses">
-            <label v-if="hasLabel" :for="inputId" :class="labelClasses">
+            <label
+                v-if="hasLabel"
+                :id="labelId"
+                :for="inputId"
+                :class="labelClasses">
                 <!--
                     @slot Override the label
                     @binding {string} label - label property 
@@ -360,7 +255,11 @@ const innerFieldClasses = defineClasses(
             </label>
         </div>
         <template v-else>
-            <label v-if="hasLabel" :for="inputId" :class="labelClasses">
+            <label
+                v-if="hasLabel"
+                :id="labelId"
+                :for="inputId"
+                :class="labelClasses">
                 <!--
                     @slot Override the label
                     @binding {string} label - label property 
