@@ -9,18 +9,24 @@ const radio = ref("small");
         <o-field>
             <o-radio
                 v-model="radio"
+                name="sizes"
                 native-value="small"
                 size="small"
                 label="Small" />
         </o-field>
 
         <o-field>
-            <o-radio v-model="radio" native-value="default" label="Default" />
+            <o-radio
+                v-model="radio"
+                name="sizes"
+                native-value="default"
+                label="Default" />
         </o-field>
 
         <o-field>
             <o-radio
                 v-model="radio"
+                name="sizes"
                 native-value="medium"
                 size="medium"
                 label="Medium" />
@@ -29,6 +35,7 @@ const radio = ref("small");
         <o-field>
             <o-radio
                 v-model="radio"
+                name="sizes"
                 native-value="large"
                 size="large"
                 label="Large" />
