@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import { computed, type PropType } from "vue";
+import { computed } from "vue";
 
 import { getDefault } from "@/utils/config";
 import { toCssDimension } from "@/utils/helpers";
 import { defineClasses } from "@/composables";
 
-import type { ComponentClass } from "@/types";
+import type { SkeletonProps } from "./props";
 
 /**
  * A placeholder for content to load
@@ -18,75 +18,17 @@ defineOptions({
     configField: "skeleton",
 });
 
-const props = defineProps({
-    /** Override existing theme classes completely */
-    override: { type: Boolean, default: undefined },
-    /** Show or hide loader	 */
-    active: { type: Boolean, default: true },
-    /** Show a loading animation */
-    animated: {
-        type: Boolean,
-        default: () => getDefault("skeleton.animated", true),
-    },
-    /** Custom width */
-    width: { type: [Number, String], default: undefined },
-    /** Custom height */
-    height: { type: [Number, String], default: undefined },
-    /** Show a circle shape */
-    circle: { type: Boolean, default: false },
-    /** Enable rounded style */
-    rounded: {
-        type: Boolean,
-        default: () => getDefault("skeleton.rounded", true),
-    },
-    /** Number of shapes to display */
-    count: { type: Number, default: 1 },
-    /**
-     * Size of skeleton
-     * @values small, medium, large
-     */
-    size: { type: String, default: undefined },
-    /**
-     * Skeleton position in relation to the element
-     * @values left, centered, right
-     */
-    position: {
-        type: String as PropType<"left" | "centered" | "right">,
-        default: "left",
-        validator: (value: string) =>
-            ["left", "centered", "right"].indexOf(value) > -1,
-    },
-    // class props (will not be displayed in the docs)
-    /** Class of the root element */
-    rootClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the skeleton animation */
-    animationClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the skeleton position */
-    positionClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the skeleton item */
-    itemClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the skeleton item rounded */
-    itemRoundedClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the skeleton size */
-    sizeClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
+const props = withDefaults(defineProps<SkeletonProps>(), {
+    override: undefined,
+    active: true,
+    animated: () => getDefault("skeleton.animated", true),
+    width: undefined,
+    height: undefined,
+    circle: false,
+    rounded: () => getDefault("skeleton.rounded", true),
+    count: 1,
+    size: undefined,
+    position: "left",
 });
 
 const itemStyle = computed(() => ({
