@@ -47,6 +47,7 @@ const itemValue = props.value || useId();
 
 const slots = useSlots();
 
+// provided data is a computed ref to enjure reactivity
 const providedData = computed<StepItemComponent<T>>(() => ({
     ...props,
     value: itemValue,
@@ -60,7 +61,7 @@ const providedData = computed<StepItemComponent<T>>(() => ({
     deactivate,
 }));
 
-// inject functionalities and data from the parent carousel component
+/** inject functionalities and data from the parent component */
 const { parent, item } = useProviderChild<StepsComponent>({
     data: providedData,
 });
@@ -175,20 +176,21 @@ const panelClasses = defineClasses(["stepPanelClass", "o-steps__panel"]);
         :appear="parent.animateInitially"
         @after-enter="afterEnter"
         @before-leave="beforeLeave">
-        <template v-if="!parent.destroyOnHide || (isActive && visible)">
-            <div
-                v-show="isActive && visible"
-                v-bind="$attrs"
-                :id="`tabpanel-${item.identifier}`"
-                :class="panelClasses"
-                :data-id="`steps-${item.identifier}`"
-                data-oruga="steps-item"
-                :aria-labelledby="`tab-${item.identifier}`"
-                aria-roledescription="item">
-                <!-- 
-                    @slot Step item content
-                -->
-                <slot>
+        <div
+            v-show="isActive && visible"
+            v-bind="$attrs"
+            :id="`tabpanel-${item.identifier}`"
+            data-oruga="steps-item"
+            :data-id="`steps-${item.identifier}`"
+            :class="panelClasses"
+            :aria-labelledby="`tab-${item.identifier}`"
+            aria-roledescription="item">
+            <!-- 
+                @slot Step item content
+                @binding {boolean} active - if item is shown 
+            -->
+            <slot :active="isActive && visible">
+                <template v-if="!parent.destroyOnHide || (isActive && visible)">
                     <!-- injected component -->
                     <component
                         :is="component"
@@ -198,8 +200,8 @@ const panelClasses = defineClasses(["stepPanelClass", "o-steps__panel"]);
 
                     <!-- default content prop -->
                     <template v-else>{{ content }}</template>
-                </slot>
-            </div>
-        </template>
+                </template>
+            </slot>
+        </div>
     </Transition>
 </template>
