@@ -1,5 +1,5 @@
 <script setup lang="ts" generic="T">
-import { computed, useAttrs, useId, useTemplateRef } from "vue";
+import { computed, useAttrs, useId, useSlots, useTemplateRef } from "vue";
 
 import { getDefault } from "@/utils/config";
 import { defineClasses, useInputHandler } from "@/composables";
@@ -23,20 +23,20 @@ defineOptions({
 const props = withDefaults(defineProps<SwitchProps<T>>(), {
     override: undefined,
     modelValue: undefined,
+    id: () => useId(),
+    label: undefined,
     variant: () => getDefault("switch.variant"),
     passiveVariant: () => getDefault("switch.passiveVariant"),
     size: () => getDefault("switch.size"),
-    label: undefined,
-    nativeValue: undefined,
-    disabled: false,
+    position: () => getDefault("switch.right", "position"),
+    rounded: () => getDefault("switch.rounded", true),
     required: false,
+    disabled: false,
     name: undefined,
+    nativeValue: undefined,
     trueValue: undefined,
     falseValue: undefined,
-    rounded: true,
-    position: "right",
     autocomplete: () => getDefault("switch.autocomplete", "off"),
-    id: () => useId(),
     useHtml5Validation: () => getDefault("useHtml5Validation", true),
 });
 
@@ -72,27 +72,29 @@ const emits = defineEmits<{
 const inputRef = useTemplateRef("inputElement");
 
 // use form input functionalities
-const { onBlur, onFocus, onInvalid, setFocus } = useInputHandler(
+const { onBlur, onFocus, onInvalid, setFocus, doClick } = useInputHandler(
     inputRef,
     emits,
     props,
 );
+
 // inject parent field component if used inside one
 const { parentField } = injectField();
 
+// set field labelId or create a unique label id if a label is given
+const labelId =
+    !!parentField.value || !!props.label || !!useSlots().label
+        ? parentField.value?.labelId || useId()
+        : undefined;
+
+// if no `label` is given and `id` is given set as `for` property on o-field wrapper
+if (!props.label && props.id) parentField.value?.setInputId(props.id);
+
 const vmodel = defineModel<T>({ default: undefined });
-
-// if not `label` is given and `id` is given set as `for` property on o-field wrapper
-if (!props.label && props.id) parentField?.value?.setInputId(props.id);
-
-const _trueValue =
-    typeof props.trueValue === "undefined" ? true : props.trueValue;
-const _falseValue =
-    typeof props.falseValue === "undefined" ? false : props.falseValue;
 
 const isChecked = computed(
     () =>
-        vmodel.value === _trueValue ||
+        vmodel.value === (props.trueValue ?? true) ||
         (Array.isArray(vmodel.value) &&
             vmodel.value.includes(props.nativeValue)),
 );
@@ -101,12 +103,17 @@ function onInput(event: Event): void {
     emits("input", vmodel.value, event);
 }
 
+function clickInput(): void {
+    setFocus();
+    doClick();
+}
+
 // --- Computed Component Classes ---
 
 const attrs = useAttrs();
 
 const inputBind = computed(() => ({
-    ...parentField?.value?.inputAttrs,
+    ...parentField.value?.inputAttrs,
     ...attrs,
 }));
 
@@ -169,14 +176,7 @@ defineExpose({ focus: setFocus, value: vmodel });
 </script>
 
 <template>
-    <label
-        ref="label"
-        :class="rootClasses"
-        data-oruga="switch"
-        role="switch"
-        :aria-checked="isChecked"
-        @click="setFocus"
-        @keydown.prevent.enter="setFocus">
+    <div :class="rootClasses" data-oruga="switch">
         <input
             v-bind="inputBind"
             :id="id"
@@ -186,28 +186,33 @@ defineExpose({ focus: setFocus, value: vmodel });
             role="switch"
             data-oruga-input="switch"
             :class="inputClasses"
-            :disabled="disabled"
-            :required="required"
             :name="name"
-            :autocomplete="autocomplete"
             :value="nativeValue"
-            :true-value="_trueValue"
-            :false-value="_falseValue"
-            @click.stop
+            :true-value="trueValue ?? true"
+            :false-value="falseValue ?? false"
+            :required="required"
+            :disabled="disabled"
+            :autocomplete="autocomplete"
+            :aria-checked="isChecked"
+            :aria-labelledby="labelId"
             @blur="onBlur"
             @focus="onFocus"
             @invalid="onInvalid"
-            @input="onInput" />
+            @change="onInput" />
 
-        <span :class="switchClasses">
+        <span :class="switchClasses" @click.prevent="clickInput">
             <span :class="switchCheckClasses"></span>
         </span>
 
-        <span v-if="label || $slots.default" :class="labelClasses">
+        <label
+            v-if="label || $slots.default"
+            :id="labelId"
+            :for="id"
+            :class="labelClasses">
             <!--
                 @slot Override the label, default is label prop 
             -->
             <slot>{{ label }}</slot>
-        </span>
-    </label>
+        </label>
+    </div>
 </template>
