@@ -39,6 +39,7 @@ import {
     useMatchMedia,
     useDebounce,
     useObjectMap,
+    useSequentialId,
 } from "@/composables";
 
 import type { ClassBind } from "@/types";
@@ -352,14 +353,15 @@ const provideData = computed<TableComponent>(() => ({
 }));
 
 /** provide functionalities and data to child item components */
-const { sortedItems } = useProviderParent<TableColumnComponent<T>>(slotRef, {
+const { childItems } = useProviderParent<TableColumnComponent<T>>({
+    rootRef: slotRef,
     data: provideData,
 });
 
 /** all defined columns */
 const tableColumns = computed<TableColumnItem<T>[]>(() => {
-    if (!sortedItems.value) return [];
-    return sortedItems.value.map((column) => ({
+    if (!childItems.value) return [];
+    return childItems.value.map((column) => ({
         index: column.index,
         identifier: column.identifier,
         ...toValue(column.data!),
@@ -368,9 +370,12 @@ const tableColumns = computed<TableColumnItem<T>[]>(() => {
     }));
 });
 
+// create a unique id sequence
+const { nextSequence } = useSequentialId();
+
 /** all defined data elements as an object map */
 const tableData = computed<TableRow<T>[]>(() =>
-    useObjectMap(props.data, props.rowKey),
+    useObjectMap(props.data, props.rowKey, nextSequence),
 );
 
 const tableRows = ref(tableData.value) as Ref<TableRow<T>[]>;
