@@ -45,6 +45,7 @@ const isHeaderUnselectable = computed(
 
 const vm = getCurrentInstance();
 
+// provided data is a computed ref to enjure reactivity
 const providedData = computed<TableColumnComponent<T>>(() => ({
     ...props,
     $el: vm!.proxy!,
@@ -54,6 +55,7 @@ const providedData = computed<TableColumnComponent<T>>(() => ({
     tdClasses: tdClasses.value,
 }));
 
+/** inject functionalities and data from the parent component */
 const { parent, item } = useProviderChild<TableComponent>({
     data: providedData,
 });
@@ -120,7 +122,7 @@ const filters = {} as Record<string, string>;
 </script>
 
 <template>
-    <span :data-id="item.identifier" data-oruga="table-column">
+    <span data-oruga="table-column" :data-id="item.identifier">
         {{ label }}
 
         <!--
