@@ -1,12 +1,5 @@
 <script setup lang="ts">
-import {
-    ref,
-    computed,
-    watch,
-    nextTick,
-    type PropType,
-    type Component,
-} from "vue";
+import { ref, computed, watch, nextTick, type Component } from "vue";
 
 import PositionWrapper from "../utils/PositionWrapper.vue";
 
@@ -18,7 +11,7 @@ import {
     useClickOutside,
 } from "@/composables";
 
-import type { ComponentClass, DynamicComponent } from "@/types";
+import type { TooltipProps } from "./props";
 
 /**
  * Display a brief helper text to your user
@@ -31,142 +24,22 @@ defineOptions({
     configField: "tooltip",
 });
 
-const props = defineProps({
-    /** Override existing theme classes completely */
-    override: { type: Boolean, default: undefined },
-    /** Whether tooltip is active or not, use v-model:active to make it two-way binding */
-    active: { type: Boolean, default: false },
-    /** Tooltip text, unnecessary when content slot is used */
-    label: { type: String, default: undefined },
-    /**
-     * Color of the tooltip
-     * @values primary, info, success, warning, danger, and any other custom color
-     */
-    variant: {
-        type: String,
-        default: () => getDefault("tooltip.variant"),
-    },
-    /**
-     * Position of the Tooltip relative to the trigger
-     * @values auto, top, bottom, left, right, top-right, top-left, bottom-left, bottom-right
-     */
-    position: {
-        type: String,
-        default: () => getDefault("tooltip.position", "auto"),
-        validator: (value: string) =>
-            [
-                "auto",
-                "top",
-                "bottom",
-                "left",
-                "right",
-                "top-right",
-                "top-left",
-                "bottom-left",
-                "bottom-right",
-            ].indexOf(value) > -1,
-    },
-    /** Tooltip will be always active */
-    always: { type: Boolean, default: false },
-    /** Tooltip will be disabled */
-    disabled: { type: Boolean, default: false },
-    /** Tooltip default animation */
-    animation: {
-        type: String,
-        default: () => getDefault("tooltip.animation", "fade"),
-    },
-    /** Tooltip will be multilined */
-    multiline: { type: Boolean, default: false },
-    /** Tooltip trigger tag name */
-    triggerTag: {
-        type: [String, Object, Function] as PropType<DynamicComponent>,
-        default: () => getDefault("tooltip.triggerTag", "div"),
-    },
-    /**
-     * Tooltip trigger events
-     * @values hover, click, focus, contextmenu
-     */
-    triggers: {
-        type: Array as PropType<string[]>,
-        default: () => getDefault("tooltip.triggers", ["hover"]),
-        validator: (values: string[]) =>
-            values.filter(
-                (value) =>
-                    ["click", "hover", "contextmenu", "focus"].indexOf(value) >
-                    -1,
-            ).length === values.length,
-    },
-    /** Tooltip delay before it appears (number in ms) */
-    delay: { type: Number, default: undefined },
-    /**
-     * Tooltip auto close options (pressing escape, clicking the content or outside)
-     * @values true, false, content, outside, escape
-     */
-    closeable: {
-        type: [Array, Boolean] as PropType<string[] | boolean>,
-        default: () =>
-            getDefault("tooltip.closeable", ["escape", "outside", "content"]),
-    },
-    /**
-     * Append the component to another part of the DOM.
-     * Set `true` to append the component to the body.
-     * In addition, any CSS selector string or an actual DOM node can be used.
-     */
-    teleport: {
-        type: [Boolean, String, Object],
-        default: () => getDefault("dropdown.teleport", false),
-    },
-    // class props (will not be displayed in the docs)
-    /** Class of the root element */
-    rootClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class when the dropdown is teleported */
-    teleportClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the tooltip content */
-    contentClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the tooltip trigger position */
-    positionClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the tooltip trigger */
-    triggerClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the tooltip content when is multiline */
-    multilineClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the tooltip trigger when is always visible */
-    alwaysClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the tooltip variant */
-    variantClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the tooltip arrow */
-    arrowClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
-    /** Class of the tooltip arrow position */
-    arrowPositionClass: {
-        type: [String, Array, Function] as PropType<ComponentClass>,
-        default: undefined,
-    },
+const props = withDefaults(defineProps<TooltipProps>(), {
+    override: undefined,
+    active: false,
+    label: undefined,
+    variant: () => getDefault("tooltip.variant"),
+    position: () => getDefault("tooltip.position", "auto"),
+    always: false,
+    disabled: false,
+    animation: () => getDefault("tooltip.animation", "fade"),
+    multiline: false,
+    triggerTag: () => getDefault("tooltip.triggerTag", "div"),
+    triggers: () => getDefault("tooltip.triggers", ["hover"]),
+    delay: undefined,
+    closeable: () =>
+        getDefault("tooltip.closeable", ["escape", "outside", "content"]),
+    teleport: () => getDefault("dropdown.teleport", false),
 });
 
 const emits = defineEmits<{
