@@ -7,8 +7,9 @@ export type ObjectMap<T> = Array<{
 
 /** wrap entities in an array of objects with key/value attributes */
 export function useObjectMap<T>(
-    values?: MaybeRefOrGetter<Array<T>>,
-    key?: string | number | symbol,
+    values: MaybeRefOrGetter<Array<T>> | undefined,
+    key: string | number | symbol | undefined,
+    uuid: () => string,
 ): ObjectMap<T> {
     if (!values || !toValue(values)?.length) return [];
     return toValue(values).map((value: T) => ({
@@ -16,7 +17,7 @@ export function useObjectMap<T>(
         key:
             // if no key is given and data is object, create unique row id for each row
             key && value && typeof value === "object"
-                ? (value[key as keyof T] as string) || crypto.randomUUID()
-                : crypto.randomUUID(),
+                ? (value[key as keyof T] as string) || uuid()
+                : uuid(),
     }));
 }
